organising oauth code into a "features" architecture
