+++ frontend/src/features/oauths/components/LoginSuccess.tsx
@@ -1,0 +1,9 @@
+const LoginSuccess = () => {
+	return (
+		<div>
+			<p>Please wait a moment...</p>
+		</div>
+	);
+};
+
+export default LoginSuccess;
